requirements.txt
# Core Dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.0.0

# Reinforcement Learning
gymnasium>=0.28.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Configuration
pyyaml>=6.0
hydra-core>=1.3.0

# Logging & Monitoring
tensorboard>=2.10.0
tqdm>=4.64.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
requests>=2.28.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
